refactor: extract shared JSON parse utilities used by all evidence parsers
parseJson and withSegmentIds were duplicated across parse-evidence-query,
parse-evidence-plan, and parse-evidence-combined. Extract to parse-json-utils
with clearer names (parseJsonFromModelOutput, injectSegmentIds) and update
all three parsers to use the shared implementation.
