+++ packages/diagnosis/src/parse-evidence-combined.ts
@@ -3,73 +3,18 @@ import {
   type EvidenceQueryRef,
   type EvidenceQueryResponse,
 } from "3am-core";
+import { parseJsonFromModelOutput, injectSegmentIds } from "./parse-json-utils.js";
 
 export type EvidenceCombinedResult =
   | { kind: "clarification"; clarificationQuestion: string }
   | { kind: "answer"; response: EvidenceQueryResponse };
-
-/**
- * Extracts JSON from model output that may contain prose and/or code fences.
- * Same strategy as parse-evidence-query.ts.
- */
-function parseJson(raw: string): unknown {
-  // Attempt 1: direct parse
-  try {
-    return JSON.parse(raw);
-  } catch {
-    // continue
-  }
-
-  // Attempt 2: first code fence
-  const fenceMatch = /```(?:json)?\s*\n([\s\S]*?)\n\s*```/.exec(raw);
-  if (fenceMatch?.[1] !== undefined) {
-    try {
-      return JSON.parse(fenceMatch[1].trim());
-    } catch {
-      // continue
-    }
-  }
-
-  // Attempt 3: first '{' to last '}'
-  const start = raw.indexOf("{");
-  const end = raw.lastIndexOf("}");
-  if (start !== -1 && end > start) {
-    try {
-      return JSON.parse(raw.slice(start, end + 1));
-    } catch {
-      // fall through
-    }
-  }
-
-  throw new Error("Failed to parse combined evidence output as JSON");
-}
-
-function withSegmentIds(parsedSegments: unknown): unknown {
-  if (!Array.isArray(parsedSegments)) return parsedSegments;
-
-  return parsedSegments.map((segment, index) => {
-    if (!segment || typeof segment !== "object" || Array.isArray(segment)) {
-      return segment;
-    }
-
-    const record = segment as Record<string, unknown>;
-    if (typeof record["id"] === "string" && record["id"].length > 0) {
-      return record;
-    }
-
-    return {
-      ...record,
-      id: `seg_${index + 1}`,
-    };
-  });
-}
 
 export function parseEvidenceCombined(
   raw: string,
   meta: { question: string },
   allowedRefs: EvidenceQueryRef[],
 ): EvidenceCombinedResult {
-  const parsed = parseJson(raw) as Record<string, unknown>;
+  const parsed = parseJsonFromModelOutput(raw) as Record<string, unknown>;
   const mode = parsed["mode"];
 
   // Shape A: clarification
@@ -89,7 +34,7 @@ export function parseEvidenceCombined(
   const withQuestion = {
     question: meta.question,
     status: parsed["status"],
-    segments: withSegmentIds(parsed["segments"] ?? []),
+    segments: injectSegmentIds(parsed["segments"] ?? []),
     evidenceSummary: { traces: 0, metrics: 0, logs: 0 },
     followups: [],
     noAnswerReason: parsed["noAnswerReason"],
